backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.21.0",
    "fastapi>=0.115.12",
    "fastapi-socketio>=0.0.10",
    "passlib[bcrypt]>=1.7.4",
    "pydantic[email]>=2.11.5",
    "python-dotenv>=1.1.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "ruff>=0.11.12",
    "sqlalchemy>=2.0.41",
    "uvicorn[standard]>=0.34.3",
    "websockets>=15.0.1",
]

[tool.ruff]
ignore = ["B008"]
